refactor(storage): remove unused FileSystemEntry.FullPath; update providers to use Name/IsDirectory/Length only
