+++ FtpServer.Core/Server/Commands/Handlers/AlloHandler.cs
@@ -1,6 +1,6 @@
 using FtpServer.Core.Protocol;
 
-namespace FtpServer.Core.Server.Commands;
+namespace FtpServer.Core.Server.Commands.Handlers;
 
 internal sealed class AlloHandler : IFtpCommandHandler
 {
