+++ FtpServer.Core/Server/Commands/Handlers/RetrHandler.cs
@@ -4,7 +4,7 @@ using FtpServer.Core.Observability;
 using FtpServer.Core.Protocol;
 using Microsoft.Extensions.Options;
 
-namespace FtpServer.Core.Server.Commands;
+namespace FtpServer.Core.Server.Commands.Handlers;
 
 internal sealed class RetrHandler : IFtpCommandHandler
 {
