+++ FtpServer.Core/Server/Commands/Handlers/RntoHandler.cs
@@ -1,7 +1,7 @@
 using FtpServer.Core.Abstractions;
 using FtpServer.Core.Protocol;
 
-namespace FtpServer.Core.Server.Commands;
+namespace FtpServer.Core.Server.Commands.Handlers;
 
 internal sealed class RntoHandler : IFtpCommandHandler
 {
