+++ FtpServer.Core/Server/Commands/Handlers/TypeHandler.cs
@@ -1,6 +1,6 @@
 using FtpServer.Core.Protocol;
 
-namespace FtpServer.Core.Server.Commands;
+namespace FtpServer.Core.Server.Commands.Handlers;
 
 internal sealed class TypeHandler : IFtpCommandHandler
 {
